refactor: remove provider
