--- frontend/packages/app/src/pages/dashboard/leadership-view/provider/context.ts
@@ -1,18 +1,0 @@
-/**
- * External dependencies.
- */
-import { createContext, useContextSelector } from "use-context-selector";
-
-export interface LeadershipViewContextProps {
-  state: {};
-  actions: {};
-}
-
-export const LeadershipViewContext = createContext<LeadershipViewContextProps>({
-  state: {},
-  actions: {},
-});
-
-export const useLeadershipView = <T>(
-  selector: (state: LeadershipViewContextProps) => T,
-) => useContextSelector(LeadershipViewContext, selector);
